Restructured build to use Ant
Have reverted back to use ant for building and deploying the artifacts
to Maven central.  This is good because it gives me more power, which I
need (for example to compile the stdlib to JVM classfiles).  At the
moment, it appears to be working although I need to figure out how to
handle the gpg signature.