Initial support for structured codegen
Instead of just writing out strings during code generation, the compiler
now generates an AST representing Java files.  This is then separately
translated into a string and written into a java file.  The advantage of
this is that it simplifies the code generator, and enables for
customisation (e.g. supporting different source levels, etc).