Support unbound operators / comparators
When performing arithmetic and/or comparisons on dynamically-sized
integer types, we cannot use Java's built-in operators.  Instead, we
have to invoke a method of some kind of the received.  For example, when
comparing to unbound Whiley ints, we use the "equals" method as these
are translated into instances of BigInteger.