Add support for DoWhile and fix some coercions
This adds support for translating do-while statements, and also fixes
some issues related to coercions.  For example, the return of an array
length is expression is type as "int" in Whiley, which should correspond
to a BigInteger.  But, in fact, the type of an arraylength bytecode in
Java in int.