Add JvmRun command
The JvmRun command allows (in principle) one to run a compiled class
file from the "wy" tool.  At this stage, the only problem is that the
std library is not on the classpath.  Potentially, this can be remedied
in the wy script.  For example, by just importing all jar files.